// packages/@raikou/system/src/core/RaikouProvider/theme.types.ts
export type RaikouColorsTuple = readonly [
  string,
  string,
  string,
  string,
  string,
  string,
  string,
  string,
  string,
  string,
  ...string[],
];

export type RaikouColorShade = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9;

export interface RaikouPrimaryShade {
  light: RaikouColorShade;
  dark: RaikouColorShade;
}

export type DefaultRaikouColor =
  | 'dark'
  | 'gray'
  | 'red'
  | 'pink'
  | 'grape'
  | 'violet'
  | 'indigo'
  | 'blue'
  | 'cyan'
  | 'green'
  | 'lime'
  | 'yellow'
  | 'orange'
  | 'teal'
  | (string & {});

export interface RaikouThemeColorsOverride {}

export type RaikouThemeColors = RaikouThemeColorsOverride extends {
  colors: Record<infer CustomColors, RaikouColorsTuple>;
}
  ? Record<CustomColors, RaikouColorsTuple>
  : Record<DefaultRaikouColor, RaikouColorsTuple>;

export type RaikouColor = keyof RaikouThemeColors;